+++ dokumentaatio/changelog.md
@@ -41,3 +41,4 @@
 - added user stats saving to database
 - added Werkzeug for user password security
 - added another gamemode and gamemode selection screen
+- moved score keeping to another class